refactor admin-actions
